chore: move UserFdw model to app/Models and remove unused UserFdwService
